# Swagger/swagg.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "MovieTicketBooking(MTB) Application",
        "description": "API document for MovieTicketBooking Service",
        "x-task1-appname": "MTB",
        "x-task1-seal": 104437,
        "version": "23.03.0"
    },
...
    "servers": [
        {
            "url": "",
            "description": "http://localhost:1111-Relative server url",
            "variables": {}
        }
    ],
    "paths": {
...
        "/theaters/{theaterId}": {
            "get": {
                "tags": [
                    "THEATER TABLE"
                ],
                "summary": "Get a theater by ID",
                "operationId": "getTheaterById",
                "parameters": [
                    {
                        "name": "theaterId",
                        "in": "path",
                        "description": "ID of the theater",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Theater found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Theater"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Theater not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "THEATER TABLE"
                ],
                "summary": "Update a theater",
                "operationId": "updateTheater",
                "parameters": [
                    {
                        "name": "theaterId",
                        "in": "path",
                        "description": "ID of the theater",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Theater"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Theater updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Theater"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Theater not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "THEATER TABLE"
                ],
                "summary": "Delete a theater",
                "operationId": "deleteTheater",
                "parameters": [
                    {
                        "name": "theaterId",
                        "in": "path",
                        "description": "ID of the theater",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Theater deleted successfully"
                    },
                    "404": {
                        "description": "Theater not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/shows": {
            "get": {
                "tags": [
                    "SHOWS TABLE"
                ],
                "summary": "Get all shows",
                "operationId": "getAllShows",
                "responses": {
                    "200": {
                        "description": "List of shows",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Show"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "SHOWS TABLE"
                ],
                "summary": "Create a new show",
                "operationId": "createShow",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Show"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Show created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Show"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/shows/{showId}": {
            "get": {
                "tags": [
                    "SHOWS TABLE"
                ],
                "summary": "Get a show by ID",
                "operationId": "getShowById",
                "parameters": [
                    {
                        "name": "showId",
                        "in": "path",
                        "description": "ID of the show",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Show"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Show not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "SHOWS TABLE"
                ],
                "summary": "Update a show",
                "operationId": "updateShow",
                "parameters": [
                    {
                        "name": "showId",
                        "in": "path",
                        "description": "ID of the show",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Show"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Show updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Show"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Show not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "SHOWS TABLE"
                ],
                "summary": "Delete a show",
                "operationId": "deleteShow",
                "parameters": [
                    {
                        "name": "showId",
                        "in": "path",
                        "description": "ID of the show",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Show deleted successfully"
                    },
                    "404": {
                        "description": "Show not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorInfo"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Theater": {
                "type": "object",
                "properties": {
                    "TheaterID": {
                        "type": "integer",
                        "example": 1
                    },
                    "Name": {
                        "type": "string",
                        "example": "Theater A"
                    },
                    "TotalSeats": {
                        "type": "string",
                        "example": "200"
                    },
                    "THEATER_CinemaID": {
                        "type": "integer",
                        "example": 1
                    }
                }
            },
            "Show": {
                "type": "object",
                "properties": {
                    "showId": {
                        "type": "integer"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "startTime": {
                        "type": "string",
                        "format": "time"
                    },
                    "endTime": {
                        "type": "string",
                        "format": "time"
                    },
                    "theaterId": {
                        "type": "integer"
                    },
                    "movieId": {
                        "type": "integer"
                    }
                },
                "required": [
                    "date",
                    "startTime",
                    "endTime",
                    "theaterId",
                    "movieId"
                ]
            },
...
            "ErrorInfo": {
                "type": "object",
                "properties": {
                    "errorCode": {
                        "type": "string"
                    },
                    "errorMessage": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "bbt": {
                "type": "apiKey",
                "in": "header",
                "name": "BBT"
            },
            "token": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization"
            }
        }
    }
}